Adjusted codes to MVC structure, modified css setting for searching input
