+++ .gitignore
@@ -1,5 +1,5 @@
 .env
-docker-compose.yaml
+app_bk.py
 test_nltk.py
 test.py
 testESquery.py
