+++ static/index.html
@@ -26,7 +26,7 @@
     <div class="nav-spacer"></div>
     <div class="search-container">
         <div class="search-container__slogan">
-            <p class="search-container__main-title">給您便利的育兒資訊</p>
+            <p class="search-container__main-title">給您快速便利的育兒資訊</p>
             <p class="search-container__subtitle">與寶寶一同成長，大人小孩都開心</p>
         </div>
         <div class="search-container__area">
