+++ static/index_styles.css
@@ -125,6 +125,8 @@ html{
 }
 
 .search-container__input {
+    max-width: 300px;
+    width: 100%;
     padding: 10px;
     font-size: 16px;
     margin-right: 10px;
@@ -152,7 +154,8 @@ html{
     position: absolute;
     top: 60px;
     left: 20px;
-    width: 300px;
+    max-width: 300px;
+    width: 100%;
     text-align: left;
     border: 2px solid #333;
     background-color: white;
